support linux arm docker image
resolve #673
