+++ .github/workflows/docker.yml
@@ -14,11 +14,15 @@ jobs:
     steps:
       - uses: actions/setup-go@v2
         with:
-          go-version: 1.18
+          go-version: 1.19
+
+      - uses: docker/setup-qemu-action@v2
+
+      - uses: docker/setup-buildx-action@v2
 
       - uses: actions/checkout@v2
 
-      - run: go run ./lib/utils/docker $GITHUB_REF
+      - run: go run ./lib/docker $GITHUB_REF
         env:
           DOCKER_TOKEN: ${{secrets.ROD_GITHUB_ROBOT}}
